requirements.txt
PySide6
mss
Pillow
httpx
rapidfuzz
winsdk
pydantic>=2.0
platformdirs

